+++ src/test/java/de/dealog/msg/TestUtils.java
@@ -3,6 +3,8 @@ package de.dealog.msg;
 import de.dealog.msg.persistence.MessageEntity;
 import de.dealog.msg.rest.MessageRest;
 
+import java.util.Date;
+
 public class TestUtils {
 
     public static MessageEntity buildMessage(final String identifier, final String headline, final String description) {
@@ -10,6 +12,7 @@ public class TestUtils {
         message.setIdentifier(identifier);
         message.setDescription(description);
         message.setHeadline(headline);
+        message.setPublishedAt(new Date());
         return message;
     }
 
@@ -18,6 +21,7 @@ public class TestUtils {
         messageRest.setIdentifier(identifier);
         messageRest.setDescription(description);
         messageRest.setHeadline(headline);
+        messageRest.setPublishedAt(new Date());
         return messageRest;
     }
 }